+++ modules/Student_Billing/StudentBalances.php
@@ -2,7 +2,15 @@
 
 DrawHeader( ProgramTitle() );
 
+$_REQUEST['cumulative_balance'] = issetVal( $_REQUEST['cumulative_balance'] );
+
 $extra['SELECT'] = ',(COALESCE((SELECT SUM(f.AMOUNT) FROM billing_fees f WHERE f.STUDENT_ID=ssm.STUDENT_ID AND f.SYEAR=ssm.SYEAR),0)-COALESCE((SELECT SUM(p.AMOUNT) FROM billing_payments p WHERE p.STUDENT_ID=ssm.STUDENT_ID AND p.SYEAR=ssm.SYEAR),0)) AS BALANCE';
+
+if ( $_REQUEST['cumulative_balance'] === 'Y' )
+{
+	// @since 10.3 Add "Cumulative Balance over school years" checkbox.
+	$extra['SELECT'] = ',(COALESCE((SELECT SUM(f.AMOUNT) FROM billing_fees f WHERE f.STUDENT_ID=ssm.STUDENT_ID),0)-COALESCE((SELECT SUM(p.AMOUNT) FROM billing_payments p WHERE p.STUDENT_ID=ssm.STUDENT_ID),0)) AS BALANCE';
+}
 
 $extra['columns_after'] = [ 'BALANCE' => _( 'Balance' ) ];
 
@@ -22,6 +30,18 @@ $extra2 = $extra;
 
 if ( $_REQUEST['search_modfunc'] === 'list' )
 {
+	// @since 10.3 Add "Cumulative Balance over school years" checkbox.
+	$cumulative_balance_onclick_url = ( $_REQUEST['cumulative_balance'] === 'Y' ?
+		PreparePHP_SELF( $_REQUEST, [], [ 'cumulative_balance' => '' ] ) :
+		PreparePHP_SELF( $_REQUEST, [], [ 'cumulative_balance' => 'Y' ] ) );
+
+	$input_cumulative_balance = '<input type="checkbox" name="cumulative_balance" value="Y" onclick="' .
+		AttrEscape( 'ajaxLink(' . json_encode( $cumulative_balance_onclick_url ) . ');' ) . '"' .
+		( $_REQUEST['cumulative_balance'] === 'Y' ? 'checked' : '' ) . ' autocomplete="off" />';
+
+	DrawHeader( '<label class="checkbox-label">' . $input_cumulative_balance . ' ' .
+		_( 'Cumulative Balance over school years' ) . '</label>' );
+
 	// Call GetStuList() only so we calculate the $total.
 	GetStuList( $extra );